Start adjusting basic architecture
Made some major changes to penumbra. Trying to get that to build.

Added a demo file with some UTF-8 text that seemed interesting,
although I've forgotten if/where I'm actually using it.

This version is quite broken, but it seemed like a pretty decent
milestone since this is where I had to stop yesterday.